+++ src/udp_class.hh
@@ -52,11 +52,11 @@ public:
   Socket();
 
   void
-  open_as_client( char const addr[], uint16_t port )
+  open_as_client( char const addr[], int port )
   { Socket_open_as_client( &data, addr, port ); }
 
   void
-  open_as_server( uint16_t port )
+  open_as_server( int port )
   { Socket_open_as_server( &data, port ); }
 
   void
